requirements.txt
aniso8601==6.0.0
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.0
cryptography==37.0.2
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
ecdsa==0.17.0
fido2==1.0.0
filelock==3.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
jsonfield==3.1.0
jsonLookup==0.9.0
pipenv==2022.6.7
platformdirs==2.5.2
psycopg==3.0.14
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.20
pyotp==2.6.0
python-decouple==3.4
python-jose==3.3.0
python-u2flib-server==5.0.1
pytz==2019.1
rsa==4.7.2
simplejson==3.17.2
six==1.12.0
SQLAlchemy==1.3.4
sqlparse==0.4.1
toml==0.10.2
typing_extensions==4.2.0
ua-parser==0.10.0
user-agents==2.2.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==0.15.5
whitenoise==5.2.0
